% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravitydatasets-package.R
\docType{data}
\name{cepii_country_information}
\alias{cepii_country_information}
\title{CEPII Country-Level Information}
\format{
A data frame with 257 rows and 8 columns:\tabular{ll}{
   variable \tab description \cr
   country_id \tab Combines a country's ISO3 code with a number identifying potential territorial transformations of the country \cr
   iso3 \tab ISO3 alphabetic \cr
   iso3num \tab ISO3 numeric \cr
   country \tab Country name \cr
   countrylong \tab Country official name \cr
   first_year \tab First year of territorial existence \cr
   last_year \tab Last year of territorial existence \cr
   countrygroup_iso3 \tab Country group (ISO3 alphabetic) \cr
   countrygroup_iso3num \tab Country group (ISO3 numeric) \cr
   iso2 \tab ISO2 alphabetic \cr
   heg_iso3_2020 \tab PENDING \cr
   heg_iso3num_2020 \tab PENDING \cr
}
}
\source{
CEPII, adapted from the World Bank and other sources
}
\description{
Allows for a full identification of each country included in the
gravity dataset and, if relevant, for a tracking of its territorial changes
(splits and merges).  Includes one observation for each territorial
configuration, mapping the full set of territorial changes that are
accounted for in gravity. For example, this dataset includes one observation
for West Germany, one for East Germany and one for the unified Germany.
}
\details{
There are differences with respect to the original Stata version.
ISO3 alphabetic codes of length zero were converted to NAs and the
attributes (i.e., column descriptions), when missing, were added after
reading the original documentation. The dynamic codes were added to the
dataset and follow from USITC's DGD codes
}
\keyword{data}
