% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravitydatasets-package.R
\docType{data}
\name{cepii_country_names}
\alias{cepii_country_names}
\title{CEPII Country-Level Information}
\format{
A data frame with 257 rows and 8 columns:\tabular{ll}{
   variable \tab description \cr
   iso3 \tab ISO3 alphabetic \cr
   iso3num \tab ISO3 numeric \cr
   country \tab Country name \cr
   countrylong \tab Country official name \cr
   first_year \tab First year of territorial existence \cr
   last_year \tab Last year of territorial existence \cr
   countrygroup_iso3 \tab Country group (ISO3 alphabetic) \cr
   countrygroup_iso3num \tab Country group (ISO3 numeric) \cr
}
}
\description{
Countries is the dataset that includes static country-level
variables, allowing for a full identification of each country included in
Gravity and, if relevant, for a tracking of its territorial changes (splits
and merges). Some of the variables provided in Countries are also included
in the main Gravity dataset.
Countries includes one observation for each territorial configuration,
mapping the full set of territorial changes that are accounted for in
Gravity. For example, Countries includes one observation for West Germany,
one for East Germany and one for the unified Germany. Similarly, it includes
one observation for Sudan before the split of South Sudan, one observation
for South Sudan, and one observation for Sudan after the split of South
Sudan.
}
\details{
There are differences with respect to the original Stata version.
ISO3 alphabetic codes of length zero were converted to NAs and the
attributes (i.e., column descriptions), when missing, were added after
reading the original documentation.
The universe of Countries (and of the Gravity dataset) is based on
CEPII's GeoDist dataset (Mayer and Zignago 2011). This dataset is augmented
with some countries and territories that either appear in the World Bank's
World Integrated Trade Solution (WITS) or that are necessary to construct
the full chain of territorial changes that have led to the creation of
countries appearing in the GeoDist dataset. In addition, some names are
updated, as well as ISO3 alphabetic numeric codes, by comparing the GeoDist
dataset with the WITS dataset and with the official source for ISO country
codes. Countries' official names also come from the WITS dataset, augmented
by Wikipedia for countries or territories that are not present in the WITS
dataset but that appear in GeoDist.
Countries (and the Gravity dataset) carefully tracks territorial changes,
i.e. the country's previous membership (in case of a split) and the
country's new membership (in case of a unification of two territories). We
only take into account the modifications that occurred over the time span
of the database, i.e 1948-2019. This is done using the CIA World Factbook
and Wikipedia.
}
\author{
CEPII, adapted from the World Bank and other sources
}
\keyword{data}
