% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravitydatasets-package.R
\docType{data}
\name{cepii_gdp_source}
\alias{cepii_gdp_source}
\title{CEPII GDP Source}
\format{
A data frame with 3 rows and 2 columns:\tabular{ll}{
   variable \tab description \cr
   gdp_source_id \tab GDP source ID \cr
   gdp_source_description \tab GDP source description \cr
}
}
\source{
CEPII
}
\description{
CEPII GDP Source
}
