% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravitydatasets-package.R
\docType{data}
\name{cepii_gravity}
\alias{cepii_gravity}
\title{CEPII Gravity}
\format{
A data frame with 4,428,288 rows and 79 columns for the period
1948-2019:\tabular{ll}{
   variable \tab description \cr
   year \tab Year \cr
   iso3_o \tab Origin ISO3 alphabetic \cr
   iso3num_o \tab Origin ISO3 numeric \cr
   iso3_o_dynamic \tab Origin DGD's dynamic country code of the country \cr
   iso3_d \tab Destination ISO3 alphabetic \cr
   iso3num_d \tab Destination ISO3 numeric \cr
   iso3_d_dynamic \tab Destination DGD's dynamic country code of the country \cr
   country_exists_o \tab 1 = Origin country exists \cr
   country_exists_d \tab 1 = Destination country exists \cr
   gmt_offset_2020_o \tab Origin GMT offset (hours) \cr
   gmt_offset_2020_d \tab Destination GMT offset (hours) \cr
   contig \tab 1 = Contiguity \cr
   dist \tab Distance between most populated cities, in km \cr
   distw \tab Population-weighted distance between most populated cities, in km \cr
   distcap \tab Distance between capitals, in km \cr
   distwces \tab Population-weighted distance between most populated cities, in km, using CES for \cr
   dist_source \tab 1 = Distance taken directly from CEPII's GeoDist; 0 = Based on close country \cr
   comlang_off \tab 1 = Common official or primary language \cr
   comlang_ethno \tab 1 = Language is spoken by at least 9\% of the population \cr
   comcol \tab 1 = Common colonizer post 1945 \cr
   comrelig \tab Common religion index \cr
   col45 \tab 1 = Pair in colonial relationship post 1945 \cr
   legal_old_o \tab Origin legal system before transition \cr
   legal_old_d \tab Destination legal system before transition \cr
   legal_new_o \tab Origin legal system after transition \cr
   legal_new_d \tab Destination legal system after transition \cr
   comleg_pretrans \tab 1 = Common legal origins before transition \cr
   comleg_posttrans \tab 1 = Common legal origins after transition \cr
   transition_legalchange \tab 1 = Common legal origin changed since transition \cr
   heg_o \tab 1 = Origin is current or former hegemon of destination \cr
   heg_d \tab 1 = Destination is current or former hegemon of origin \cr
   col_dep_ever \tab 1 = Pair ever in colonial or dependency relationship \cr
   col_dep \tab 1 = Pair currently in colonial or dependency relationship \cr
   col_dep_end_year \tab Independence date, if col_dep = 1 \cr
   col_dep_end_conflict \tab 1 = Independence involved conflict, if col_dep_ever = 1 \cr
   empire \tab Hegemon if sibling = 1 and year < sever_year \cr
   sibling_ever \tab 1 = Pair ever in sibling relationship \cr
   sibling \tab 1 = Pair currently in sibling relationship \cr
   sever_year \tab Severance year for pairs if sibling == 1 \cr
   sib_conflict \tab 1 = Pair ever in sibling relationship and conflict with hegemon \cr
   pop_o \tab Origin Population, total in thousands \cr
   pop_d \tab Destination Population, total in thousands \cr
   gdp_o \tab Origin GDP (current thousands US$) \cr
   gdp_d \tab Destination GDP (current thousands US$) \cr
   gdpcap_o \tab Origin GDP per cap (current thousands US$) \cr
   gdpcap_d \tab Destination GDP per cap (current thousands US$) \cr
   pop_source_o \tab Origin Population source \cr
   pop_source_d \tab Destination Population source \cr
   gdp_source_o \tab Origin GDP source \cr
   gdp_source_d \tab Destination GDP source \cr
   gdp_ppp_o \tab Origin GDP, PPP (current thousands international $) \cr
   gdp_ppp_d \tab Destination GDP, PPP (current thousands international $) \cr
   gdpcap_ppp_o \tab Origin GDP per cap, PPP (current thousands international $) \cr
   gdpcap_ppp_d \tab Destination GDP per cap, PPP (current thousands international $) \cr
   pop_pwt_o \tab Origin Population, total in thousands (PWT) \cr
   pop_pwt_d \tab Destination Population, total in thousands (PWT) \cr
   gdp_ppp_pwt_o \tab Origin GDP, current PPP (2011 thousands US$) (PWT) \cr
   gdp_ppp_pwt_d \tab Destination GDP, current PPP (2011 thousands US$) (PWT) \cr
   gatt_o \tab Origin GATT membership \cr
   gatt_d \tab Destination GATT membership \cr
   wto_o \tab Origin WTO membership \cr
   wto_d \tab Destination WTO membership \cr
   eu_o \tab 1 = Origin is a EU member \cr
   eu_d \tab 1 = Destination is a EU member \cr
   rta \tab 1 = RTA (source: WTO) \cr
   rta_coverage \tab Coverage of RTA (source: WTO) \cr
   rta_type \tab Type of RTA (source: WTO) \cr
   entry_cost_o \tab Origin Cost of business start-up procedures (\% of GNI per capita) \cr
   entry_cost_d \tab Destination Cost of business start-up procedures (\% of GNI per capita) \cr
   entry_proc_o \tab Origin Start-up procedures to register a business (number) \cr
   entry_proc_d \tab Destination Start-up procedures to register a business (number) \cr
   entry_time_o \tab Origin Time required to start a business (days) \cr
   entry_time_d \tab Destination Time required to start a business (days) \cr
   entry_tp_o \tab Origin Days + procedures to start a business \cr
   entry_tp_d \tab Destination Days + procedures to start a business \cr
   tradeflow_comtrade_o \tab Trade flows as reported by the origin, 1000 Current USD (source: UNSD) \cr
   tradeflow_comtrade_d \tab Trade flows as reported by the destination, 1000 Current USD (source: UNSD) \cr
   tradeflow_baci \tab Trade flow, 1000 USD (source: BACI) \cr
   manuf_tradeflow_baci \tab Trade flow of manufactured goods, 1000 USD (source: BACI) \cr
   tradeflow_imf_o \tab Trade flows as reported by the origin, 1000 Current USD (source: IMF) \cr
   tradeflow_imf_d \tab Trade flows as reported by the destination, 1000 Current USD (source: IMF) \cr
}
}
\source{
CEPII, adapted from the World Bank and other sources
}
\description{
Each observation is uniquely identified by the
combination of the ISO-3 code of the origin country, the ISO-3 code of the
destination country and the year. Country pair appears every year, even if
one of the countries actually does not exist. However, based on the
territorial changes tracked in the countries dataset, we set to missing all
variables for country pairs in which at least one of the countries does not
exist in a given year. Furthermore, we provide two dummy variables
indicating whether the origin and the destination countries exist. These
dummies allow users wishing drop non-existing country pairs from the
dataset.
A few caveats on the identification of countries through country_id must be
noted. Firstly, when countries merge, it is the new country or territorial
configuration that exists during transition year but not the old country or
territorial configuration. As an example DEU.1 (West Germany) has 1989 as
last year, not 1990, while DEU.2 (the unified Germany) has 1990 as first
year. This is consistent with the construction of underlying variables that
varies over time, such as GDP, population, trade. Secondly, since the
dataset is square in terms of country_id, there exist cases in which two
configurations of the same alphabetic ISO3 code appear bilaterally, e.g.
DEU.1 and DEU.2. While DEU.1 and DEU.2 never existed simultaneously, we
still keep these null observations to ensure that the final dataset is
square.
}
\details{
The details are the same as for the countries dataset.
}
\keyword{data}
