% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravitydatasets-package.R
\docType{package}
\name{gravitydatasets-package}
\alias{gravitydatasets}
\alias{gravitydatasets-package}
\title{gravitydatasets: Aims at gathering in a single place a set of variables adapted both from the Center for Prospective Studies and International Information (CEPII) and the US International Trade Commission (USITC) that could be useful to researchers or practitioners willing to understand the determinants of international trade. Each observation corresponds to a combination of exporter-importer-year (i.e. origin-destination-year), for which we provide trade flows, as well as geographic, cultural, trade facilitation and macroeconomic variables. The tables provided here were reshaped under Cobb's Normal Form, and in some cases contain corrections that we explain in the README. The specific sources are the Gravity database from CEPII and the International Trade and Production Database for Estimation (ITPD-E) and the Dynamic Gravity Database (DGD) from the USITC.}
\description{
Provides consistent data on international and domestic trade for 243 countries, 170 industries, and 17 years alongside variables for gravity estimation. The data sources are ITPD-E and DGD from the USITC, which use administrative data and intentionally does not include information estimated by statistical techniques, which makes the datasets suitable for gravity estimation.

Provides consistent data on international and domestic trade for 243 countries, 170 industries, and 17 years alongside variables for gravity estimation. The data sources are ITPD-E and DGD from the USITC, which use administrative data and intentionally does not include information estimated by statistical techniques, which makes the datasets suitable for gravity estimation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pacha.dev/gravitydatasets}
  \item Report bugs at \url{https://github.com/pachadotdev/gravitydatasets/issues}
}


Useful links:
\itemize{
  \item \url{https://pacha.dev/gravitydatasets}
  \item Report bugs at \url{https://github.com/pachadotdev/gravitydatasets/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Other contributors:
\itemize{
  \item  Center for Prospective Studies and International Information(CEPII) [data contributor]
  \item  United States International Trade Comission (USITC) [data contributor]
}

}
\keyword{internal}
