% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{gravitydatasets_connect}
\alias{gravitydatasets_connect}
\title{Connect to the gravitydatasets database}
\usage{
gravitydatasets_connect(dir = gravitydatasets_path())
}
\arguments{
\item{dir}{Database location on disk. Defaults to the \code{gravitydatasets}
directory inside the R user folder or the \code{gravitydatasets_DIR} environment variable
if specified.}
}
\description{
Returns a local database connection. This corresponds to a DBI-compatible
DuckDB database.
}
\examples{
\dontrun{
 DBI::dbListTables(gravitydatasets_connect())

 DBI::dbGetQuery(
  gravitydatasets_connect(),
  'SELECT * FROM gravitydatasets WHERE year = 2010'
 )
}
}
