% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{gravitydatasets_delete}
\alias{gravitydatasets_delete}
\title{Delete the gravitydatasets database from your computer}
\usage{
gravitydatasets_delete(ask = TRUE)
}
\arguments{
\item{ask}{If so, a menu will be displayed to confirm the action to
delete any existing gravitydatasets database. By default it is true.}
}
\description{
Deletes the \code{gravitydatasets} directory and all of its contents, including
all versions of the gravitydatasets database created with any DuckDB version.
}
\examples{
\dontrun{ gravitydatasets_delete() }
}
