% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{gravitydatasets_disconnect}
\alias{gravitydatasets_disconnect}
\title{Disconnect the gravitydatasets database}
\usage{
gravitydatasets_disconnect()
}
\description{
An auxiliary function to disconnect from the database.
}
\examples{
gravitydatasets_disconnect()
}
