% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{gravitydatasets_download}
\alias{gravitydatasets_download}
\title{Download the Gravity Datasets to your computer}
\usage{
gravitydatasets_download(ver = NULL, local_file = NULL)
}
\arguments{
\item{ver}{The version to download. By default it is the latest version
available on GitHub. All versions can be viewed at
\url{https://github.com/pachadotdev/gravitydatasets/releases}.}

\item{local_file}{The path to a local zip file. If this is provided, the
function will not download the file from GitHub and will instead use the
local file.}
}
\description{
This command downloads the entire database as a single zip file that is
unzipped to create the local database. If you do not want to download the
database on your home computer, run usethis::edit_r_environ() to create the
environment variable gravitydatasets_DIR with the path.
}
\examples{
\dontrun{ gravitydatasets_download() }
}
