% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravitydatasets-package.R
\docType{data}
\name{usitc_country_names}
\alias{usitc_country_names}
\title{Country names: Auxiliary table}
\format{
A data frame with 267 rows and 3 columns:
|Variable name         |Variable description                                                                                                |
|country_iso3          |ISO 3-letter alpha code of the country                                                                              |
|country_dynamic_code  |DGD's dynamic country code of the country                                                                           |
|country_name          |Name of the country                                                                                                 |
}
\source{
Adapted from the original ITPD-E dataset
}
\description{
Country names: Auxiliary table
}
\author{
Own creation
}
\keyword{data}
