% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravitydatasets-package.R
\docType{data}
\name{usitc_gravity}
\alias{usitc_gravity}
\title{USITC Gravity: Macroeconomic, geographic and institutional variables table.}
\format{
A data frame with 1,940,681 rows and 67 columns for the period 1986-2020:\tabular{ll}{
   Variable name \tab Variable description \cr
   year \tab Year of observation \cr
   iso3_o \tab 3-digit ISO code of origin country \cr
   dynamic_code_o \tab Year appropriate 3-digit code of origin country \cr
   iso3_d \tab 3-digit ISO code of destination country \cr
   dynamic_code_d \tab Year appropriate 3-digit code of destination country \cr
   colony_of_destination_ever \tab Origin country was ever a colony of the destination country \cr
   colony_of_origin_ever \tab Destination country was ever a colony of the origin country \cr
   colony_ever \tab Country pair has been in a colonial relationship \cr
   common_colonizer \tab Country pair has been colonized by a common colonizer \cr
   common_legal_origin \tab Country pair shares common legal origin \cr
   contiguity \tab Country pair shares a common border \cr
   distance \tab Population weighted distance between country pair \cr
   member_gatt_o \tab Origin country is a General Agreement on Tariffs and Trade member \cr
   member_wto_o \tab Origin country is a World Trade Organization member \cr
   member_eu_o \tab Origin country is a European Union member \cr
   member_gatt_d \tab Destination country is a General Agreement on Tariffs and Trade member \cr
   member_wto_d \tab Destination country is a World Trade Organization member \cr
   member_eu_d \tab Destination country is a European Union member \cr
   member_gatt_joint \tab Country pair are both members of the General Agreement on Tariffs and Trade \cr
   member_wto_joint \tab Country pair are both members of the World Trade Organization \cr
   member_eu_joint \tab Country pair are both members of the European Union \cr
   lat_o \tab Latitude coordinate of origin country \cr
   lng_o \tab Longitude coordinate of origin country \cr
   lat_d \tab Latitude coordinate of destination country \cr
   lng_d \tab Longitude coordinate of destination country \cr
   landlocked_o \tab Origin country is landlocked \cr
   island_o \tab Origin country is an island \cr
   region_id_o \tab Geographic region of origin country \cr
   landlocked_d \tab Destination country is landlocked \cr
   island_d \tab Destination country is an island \cr
   region_id_d \tab Geographic region of destination country \cr
   agree_pta_goods \tab Country pair is in at least one active preferential trade agreement covering goods \cr
   agree_pta_services \tab Country pair is in at least one active preferential trade agreement covering services \cr
   agree_fta \tab Country pair is in at least one free trade agreement \cr
   agree_eia \tab Country pair is in at least one economic integration agreement \cr
   agree_cu \tab Country pair is in at least one customs union \cr
   agree_psa \tab Country pair is in at least one partial scope agreement \cr
   agree_fta_eia \tab Country pair is in at least one free trade agreement and at least one economic integration agreement \cr
   agree_cu_eia \tab Country pair is in at least one customs union and at least one economic integration agreement \cr
   agree_pta \tab Country pair is in at least one active preferential trade agreement covering goods \cr
   capital_const_d \tab Capital stock at constant prices of destination country \cr
   capital_const_o \tab Capital stock at constant prices of origin country \cr
   capital_cur_d \tab Capital stock at current PPP of destination country \cr
   capital_cur_o \tab Capital stock at current PPP of origin country \cr
   gdp_pwt_const_d \tab Real, inflation-adjusted, PPP-adjusted GDP of destination country (PWT) \cr
   gdp_pwt_const_o \tab Real, inflation-adjusted, PPP-adjusted GDP of origin country (PWT) \cr
   gdp_pwt_cur_d \tab Real, current, PPP-adjusted GDP of destination country (PWT) \cr
   gdp_pwt_cur_o \tab Real, current, PPP-adjusted GDP of origin country (PWT) \cr
   pop_d \tab Population of destination country \cr
   pop_o \tab Population of origin country \cr
   hostility_level_o \tab Level of the origin/destination country’s hostility toward the destination country \cr
   hostility_level_d \tab Level of the origin/destination country’s hostility toward the origin country \cr
   common_language \tab Residents of country pair speak at least one common language \cr
   polity_o \tab Polity (political stability) score of origin country \cr
   polity_d \tab Polity (political stability) score of destination country \cr
   sanction_threat \tab There exists a threat of sanction between one country in a record towards the other \cr
   sanction_threat_trade \tab There exists a threat of trade sanction between one country in a record towards the other \cr
   sanction_imposition \tab There exists a sanction between one country in a record towards the other \cr
   sanction_imposition_trade \tab There exists a trade sanction between one country in a record towards the other \cr
   gdp_wdi_cur_o \tab Nominal GDP of origin country (WDI) \cr
   gdp_wdi_cap_cur_o \tab Nominal GDP per capita of origin country (WDI) \cr
   gdp_wdi_const_o \tab Real, current, PPP-adjusted GDP of origin country (PWT) \cr
   gdp_wdi_cap_const_o \tab Real GDP per capita of origin country (WDI) \cr
   gdp_wdi_cur_d \tab Nominal GDP of destination country (WDI) \cr
   gdp_wdi_cap_cur_d \tab Nominal GDP per capita of destination country (WDI) \cr
   gdp_wdi_const_d \tab Real, current, PPP-adjusted GDP of destination country (PWT) \cr
   gdp_wdi_cap_const_d \tab Real GDP per capita of destination country (WDI) \cr
}
}
\description{
The data goes back to 1986 and is suited to join with the \code{trade} table.
}
\details{
There are differences with respect to the original CSV files. This version provides a more compact representation of the data, with the following changes:
\itemize{
\item{Starts in 1986 instead of 1948.}
\item{Is limited to ISO codes contained in the \code{trade} table.}
\item{The \code{region_origin} and \code{region_destination} columns are provided in the \verb{region names} table as \code{region_name} and can be joined by using the \code{region_id_} columns.}
}
}
\author{
USITC, adapted from WTO, UN COMTRADE, National Geographic and other sources, with corrections made for the package
}
\keyword{data}
