% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravitydatasets-package.R
\docType{data}
\name{usitc_industry_names}
\alias{usitc_industry_names}
\title{USITC-Derived Industry Names}
\format{
A data frame with 170 rows and 2 columns:\tabular{ll}{
   Variable name \tab Variable description \cr
   industry_id \tab ITPD industry code \cr
   industry_descr \tab ITPD industry description \cr
}
}
\source{
Own creation, adapted from the original ITPD-E
}
\description{
USITC-Derived Industry Names
}
\keyword{data}
