% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravitydatasets-package.R
\docType{data}
\name{usitc_sector_names}
\alias{usitc_sector_names}
\title{USTIC-Derived Sector Names}
\format{
A data frame with 4 rows and 2 columns:\tabular{ll}{
   Variable name \tab Variable description \cr
   broad_sector_id \tab Broad sector code \cr
   broad_sector \tab Broad sector \cr
}
}
\description{
USTIC-Derived Sector Names
}
\author{
Own creation, adapted from the original ITPD-E
}
\keyword{data}
