% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravitydatasets-package.R
\docType{data}
\name{usitc_trade}
\alias{usitc_trade}
\title{USITC Trade at Sector-Level}
\format{
A data frame with 72,534,869 rows and 10 columns for the period
1986-2020:\tabular{ll}{
   Variable name \tab Variable description \cr
   exporter_iso3 \tab ISO 3-letter alpha code of the exporter \cr
   exporter_dynamic_code \tab DGD's dynamic code of the exporter \cr
   exporter_name \tab Name of the exporter \cr
   importer_iso3 \tab ISO 3-letter alpha code of the importer \cr
   importer_dynamic_code \tab DGD's dynamic code of the importer \cr
   importer_name \tab Name of the importer \cr
   year \tab Year \cr
   industry_id \tab ITPD industry code \cr
   industry_descr \tab ITPD industry description \cr
   broad_sector \tab Broad sector \cr
   trade \tab Trade flows in million of current US dollars \cr
   flag_mirror \tab Flag indicator, 1 if trade mirror value is used \cr
   flag_zero \tab Flag indicator: \code{p} if positive trade, \code{r} if the raw data contained zero and `u`` missing (unknown, assigned zero) \cr
}
}
\description{
The data goes back to 1986 for Agriculture, and to 1988 for Mining & Energy and Manufacturing. Due to administrative data limitations, the data for Services is not available before to the year 2000.
}
\details{
There are differences with respect to the original CSV file. This version provides a more compact representation of the data, with the following changes:
\itemize{
\item{The \code{exporter_name} and \code{importer_name} columns are provided in the \code{country_names} table as \code{country_name} and can be joined by using the \verb{_iso3} and \verb{_dynamic} columns.}
\item{The \code{industry_descr} column is provided in the industry names table.}
\item{The \code{broad_sector} column is provided in the sector names table and \code{broad_sector_id} was created for this version of the table.}
}
}
\author{
USITC, adapted from UN COMTRADE and other sources
}
\keyword{data}
