% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravitydatasets-package.R
\docType{data}
\name{wto_country_names}
\alias{wto_country_names}
\title{WTO Country Names}
\format{
A data frame with 232 rows and 3 columns:\tabular{ll}{
   variable \tab description \cr
   country_iso3 \tab Country ISO3 alphabetic \cr
   country_name \tab Country name \cr
}

A data frame with 972,692 rows and 5 columns for the period
1980-2016:\tabular{ll}{
   variable \tab description \cr
   pair_id \tab Exporter-Importer index \cr
   year \tab Year of observation \cr
   exporter_iso3 \tab Exporter ISO3 alphabetic \cr
   importer_iso3 \tab Importer ISO3 alphabetic \cr
   trade \tab Trade flows in current US dollars \cr
}
}
\source{
Own creation, adapted from the original SGD

WTO
}
\description{
WTO Country Names

WTO Trade
}
