% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravitydatasets-package.R
\docType{data}
\name{wto_trade}
\alias{wto_trade}
\title{WTO Trade}
\format{
A data frame with 972,692 rows and 5 columns for the period
1980-2016:\tabular{ll}{
   variable \tab description \cr
   pair_id \tab Exporter-Importer index \cr
   year \tab Year of observation \cr
   exporter_iso3 \tab Exporter ISO3 alphabetic \cr
   importer_iso3 \tab Importer ISO3 alphabetic \cr
   trade \tab Trade flows in current US dollars \cr
}
}
\source{
WTO
}
\description{
WTO Trade
}
