% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravitydatasets-package.R
\docType{package}
\name{gravitydatasets-package}
\alias{gravitydatasets}
\alias{gravitydatasets-package}
\title{gravitydatasets: Datasets for Structural Gravity Modeling}
\description{
Aims at gathering in a single place a set of variables that could be useful to researchers or practitioners that use the structural gravity model of trade. We provide trade flows, as well as geographic, cultural, trade facilitation and macroeconomic variables. The tables provided here were reshaped under Cobb's Normal Form, and in some cases contain corrections that we explain in the GitHub repository. The specific sources are the Gravity database from the Center for Prospective Studies and International Information (CEPII), the International Trade and Production Database for Estimation (ITPD-E) and the Dynamic Gravity Database (DGD) from the US International Trade Commission (USITC), and the Structural Gravity Database (SGD) from the World Trade Organization (WTO).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pacha.dev/gravitydatasets}
  \item Report bugs at \url{https://github.com/pachadotdev/gravitydatasets/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Other contributors:
\itemize{
  \item Victor Falkenheim [thesis advisor]
  \item  Center for Prospective Studies and International Information(CEPII) [data contributor]
  \item  United States International Trade Comission (USITC) [data contributor]
}

}
\keyword{internal}
