% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravitydatasets-package.R
\docType{data}
\name{usitc_country_names}
\alias{usitc_country_names}
\title{USITC-Derived Country Names}
\format{
A data frame with 267 rows and 3 columns:\tabular{ll}{
   Variable name \tab Variable description \cr
   country_iso3 \tab ISO 3-letter alpha code of the country \cr
   country_dynamic_code \tab DGD's dynamic country code of the country \cr
   country_name \tab Name of the country \cr
}
}
\source{
Own creation, adapted from the original DGD
}
\description{
USITC-Derived Country Names
}
\keyword{data}
