% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravitydatasets-package.R
\docType{data}
\name{usitc_industry_names}
\alias{usitc_industry_names}
\title{Industry names: Auxiliary table}
\format{
A data frame with 170 rows and 2 columns:
|Variable name         |Variable description                                                                                                |
|industry_id           |ITPD industry code                                                                                                  |
|industry_descr        |ITPD industry description                                                                                           |
}
\source{
Adapted from the original ITPD-E dataset
}
\description{
Industry names: Auxiliary table
}
\author{
Own creation, adapted from the original ITPD-E dataset
}
\keyword{data}
