% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravitydatasets-package.R
\docType{data}
\name{usitc_region_names}
\alias{usitc_region_names}
\title{USITC-Derived Region Names}
\format{
A data frame with 15 rows and 2 columns:\tabular{ll}{
   Variable name \tab Variable description \cr
   region_id \tab Region code \cr
   region \tab Region name \cr
}
}
\description{
USITC-Derived Region Names
}
\author{
Own creation, adapted from the original DGD
}
\keyword{data}
