% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravitydatasets-package.R
\docType{data}
\name{usitc_sector_names}
\alias{usitc_sector_names}
\title{Sector names: Auxiliary table}
\format{
A data frame with 4 rows and 2 columns:
|Variable name         |Variable description                                                                                                |
|broad_sector_id       |Broad sector code                                                                                                   |
|broad_sector          |Broad sector                                                                                                        |
}
\source{
Adapted from the original ITPD-E dataset
}
\description{
Sector names: Auxiliary table
}
\author{
Own creation, adapted from the original ITPD-E dataset
}
\keyword{data}
