% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravitydatasets-package.R
\docType{data}
\name{wto_country_names}
\alias{wto_country_names}
\title{WTO Country Names}
\format{
A data frame with 232 rows and 3 columns:\tabular{ll}{
   variable \tab description \cr
   country_iso3 \tab Country ISO3 alphabetic \cr
   country_name \tab Country name \cr
}
}
\source{
Own creation, adapted from the original SGD
}
\description{
WTO Country Names
}
